+++ CHANGELOG.md
@@ -117,6 +117,7 @@
 - Prepare the theme for localization using internationalization best practices.
 - Remove `bu_search_form_query_attributes` filter function. This is now the
  default behavior in BU-CMS.
+- Remove Content Banner support.
 - Add Grunt time reporting.
  
 ## 1.5.6